0.0.5

+++ integra-portico-business/src/main/java/xeredi/integra/model/proceso/bo/ProcesoBO.java
@@ -130,7 +130,7 @@ public class ProcesoBO {
                 arin.setId(igBO.nextVal(IgBO.SQ_INTEGRA));
                 arin.setNombre(fileEntrada.getName());
                 arin.setSentido(ArchivoSentido.E);
-                arin.setTamanio(fileEntrada.length());
+                arin.setTamanio((int) fileEntrada.length());
                 arin.setFalta(Calendar.getInstance().getTime());
 
                 arch.setArin(arin);
@@ -265,7 +265,7 @@ public class ProcesoBO {
                 arin.setId(igBO.nextVal(IgBO.SQ_INTEGRA));
                 arin.setNombre(fileSalida.getName());
                 arin.setSentido(ArchivoSentido.S);
-                arin.setTamanio(fileSalida.length());
+                arin.setTamanio((int) fileSalida.length());
                 arin.setFalta(Calendar.getInstance().getTime());
 
                 arch.setArin(arin);
